+++ cmake/share/StandaloneC.cmake
@@ -46,7 +46,7 @@ if( STANDALONE)
    endif()
 
    if( NOT STANDALONE_DEFINITIONS)
-      set( STANDALONE_DEFINITIONS ${MULLE_SCION_DEFINITIONS})
+      set( STANDALONE_DEFINITIONS ${MULLE__SCION_DEFINITIONS})
    endif()
 
    #
@@ -88,7 +88,7 @@ if( STANDALONE)
       if( NOT STANDALONE_SOURCES)
          message( FATAL_ERROR "You need to define STANDALONE_SOURCES. Add a file
 ${STANDALONE_LIBRARY_NAME}.c with contents like this to it:
-int  ___mulle_scion_unused__;
+int  ___mulle__scion_unused__;
 and everybody will be happy")
       endif()